+++ cvat-ui/src/components/cvat-app.tsx
@@ -56,6 +56,8 @@ import WebhooksPage from 'components/webhooks-page/webhooks-page';
 import CreateWebhookPage from 'components/setup-webhook-pages/create-webhook-page';
 import UpdateWebhookPage from 'components/setup-webhook-pages/update-webhook-page';
 
+import GuidePage from 'components/md-guide/guide-page';
+
 import AnnotationPageContainer from 'containers/annotation-page/annotation-page';
 import { getCore } from 'cvat-core-wrapper';
 import GlobalHotKeys, { KeyMap } from 'utils/mousetrap-react';
@@ -73,6 +75,8 @@ import EmailConfirmationPage from './email-confirmation-pages/email-confirmed';
 import EmailVerificationSentPage from './email-confirmation-pages/email-verification-sent';
 import IncorrectEmailConfirmationPage from './email-confirmation-pages/incorrect-email-confirmation';
 import CreateModelPage from './create-model-page/create-model-page';
+import CreateJobPage from './create-job-page/create-job-page';
+import TaskAnalyticsPage from './task-analytics-page/task-analytics-page';
 import OrganizationWatcher from './watchers/organization-watcher';
 
 interface CVATAppProps {
@@ -458,9 +462,13 @@ class CVATApplication extends React.PureComponent<CVATAppProps & RouteComponentP
                                             <Route exact path='/projects/create' component={CreateProjectPageComponent} />
                                             <Route exact path='/projects/:id' component={ProjectPageComponent} />
                                             <Route exact path='/projects/:id/webhooks' component={WebhooksPage} />
+                                            <Route exact path='/projects/:id/guide' component={GuidePage} />
                                             <Route exact path='/tasks' component={TasksPageContainer} />
                                             <Route exact path='/tasks/create' component={CreateTaskPageContainer} />
                                             <Route exact path='/tasks/:id' component={TaskPageComponent} />
+                                            <Route exact path='/tasks/:id/analytics' component={TaskAnalyticsPage} />
+                                            <Route exact path='/tasks/:id/jobs/create' component={CreateJobPage} />
+                                            <Route exact path='/tasks/:id/guide' component={GuidePage} />
                                             <Route exact path='/tasks/:tid/jobs/:jid' component={AnnotationPageContainer} />
                                             <Route exact path='/jobs' component={JobsPageComponent} />
                                             <Route exact path='/cloudstorages' component={CloudStoragesPageComponent} />