+++ cvat-ui/src/components/project-page/project-page.tsx
@@ -60,7 +60,7 @@ export default function ProjectPageComponent(): JSX.Element {
     const [updatingProject, setUpdatingProject] = useState(false);
     const mounted = useRef(false);
 
-    const taskNamePlugins = useSelector((state: CombinedState) => state.plugins.components.taskItem.name);
+    const ribbonPlugins = useSelector((state: CombinedState) => state.plugins.components.taskItem.ribbon);
     const deletes = useSelector((state: CombinedState) => state.projects.activities.deletes);
     const taskDeletes = useSelector((state: CombinedState) => state.tasks.activities.deletes);
     const tasksActiveInferences = useSelector((state: CombinedState) => state.models.inferences);
@@ -153,10 +153,10 @@ export default function ProjectPageComponent(): JSX.Element {
                         .map((task: Task) => (
                             <TaskItem
                                 key={task.id}
+                                ribbonPlugins={ribbonPlugins}
                                 deleted={task.id in taskDeletes ? taskDeletes[task.id] : false}
                                 hidden={false}
                                 activeInference={tasksActiveInferences[task.id] || null}
-                                taskNamePlugins={taskNamePlugins}
                                 cancelAutoAnnotation={() => {
                                     dispatch(cancelInferenceAsync(task.id));
                                 }}