+++ cvat/settings/base.py
@@ -432,6 +432,9 @@ os.makedirs(TASKS_ROOT, exist_ok=True)
 PROJECTS_ROOT = os.path.join(DATA_ROOT, 'projects')
 os.makedirs(PROJECTS_ROOT, exist_ok=True)
 
+ASSETS_ROOT = os.path.join(DATA_ROOT, 'assets')
+os.makedirs(ASSETS_ROOT, exist_ok=True)
+
 SHARE_ROOT = os.path.join(BASE_DIR, 'share')
 os.makedirs(SHARE_ROOT, exist_ok=True)
 
@@ -522,9 +525,6 @@ if os.getenv('DJANGO_LOG_SERVER_HOST'):
 DATA_UPLOAD_MAX_MEMORY_SIZE = 100 * 1024 * 1024  # 100 MB
 DATA_UPLOAD_MAX_NUMBER_FIELDS = None   # this django check disabled
 DATA_UPLOAD_MAX_NUMBER_FILES = None
-LOCAL_LOAD_MAX_FILES_COUNT = 500
-LOCAL_LOAD_MAX_FILES_SIZE = 512 * 1024 * 1024  # 512 MB
-CLOUD_STORAGE_MAX_FILES_COUNT = LOCAL_LOAD_MAX_FILES_COUNT
 
 RESTRICTIONS = {
     # allow access to analytics component to users with business role
@@ -686,3 +686,7 @@ BUCKET_CONTENT_MAX_PAGE_SIZE =  500
 IMPORT_CACHE_FAILED_TTL = timedelta(days=90)
 IMPORT_CACHE_SUCCESS_TTL = timedelta(hours=1)
 IMPORT_CACHE_CLEAN_DELAY = timedelta(hours=2)
+
+ASSET_MAX_SIZE_MB = 2
+ASSET_SUPPORTED_TYPES = ('image/jpeg', 'image/png', 'image/webp', 'image/gif', 'application/pdf', )
+ASSET_MAX_COUNT_PER_GUIDE = 10